refactor(assets): Move generation logic out of internal.ts so we don't import Node-isms when we don't need them
